backend/requirements.txt
openai
gTTS
python-dotenv
SpeechRecognition
pyaudio
grpcio
grpcio-tools
langchain
